fix mtry default value
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/party@96 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ party/R/Classes.R
@@ -124,7 +124,7 @@ setClass(Class = "GlobalTestControl",
             levels = c("Bonferroni", "MonteCarlo", "Aggregated", "Raw")),
         nresample = as.integer(9999),
         randomsplits = FALSE,
-        mtry = as.integer(3),
+        mtry = as.integer(0),
         mincriterion = 0.95
     )
 )
